requirements.txt

flask
pillow
